Functionality complete
I did a ton of stuff. I'm gonna try and finish the flight tracker now, but we're very close to done. the Games and Movies pages have to be adjusted to fit the new templating so i'll leave that to you to decide how you wanna do it, most of it will likely be creative decision making.
Functionality is all there